Extract out shared methods to a module
Extracted out the shared methods in various request classes to the
`TranslationSource` module. Also pulled out the `source=` method into
the module.
