+++ lib/zendesk2/help_center/update_help_center_translation.rb
@@ -1,5 +1,6 @@
 class Zendesk2::UpdateHelpCenterTranslation
   include Zendesk2::Request
+  include Zendesk2::HelpCenter::TranslationSource::Request
 
   request_method :put
   request_body { |r| { "translation" => r.translation_params } }
@@ -9,31 +10,7 @@ class Zendesk2::UpdateHelpCenterTranslation
     @_translation_params ||= Cistern::Hash.slice(params.fetch("translation"), *Zendesk2::CreateHelpCenterTranslation.accepted_attributes)
   end
 
-  def source_id
-    Integer(params.fetch("translation").fetch("source_id"))
-  end
-
-  def source_type
-    params.fetch("translation").fetch("source_type")
-  end
-
-  def source_type_url
-    case source_type
-    when "Article"
-      "articles"
-    when "Section"
-      "sections"
-    when "Category"
-      "categories"
-    end
-  end
-
-  def locale
-    params.fetch("translation").fetch("locale") || "en-us"
-  end
-
   def mock
-    key = [source_type, source_id, locale].join("-").each_byte.inject(0) {|char, acc| acc + char }
-    mock_response("translation" => self.find!(:help_center_translations, key).merge!(translation_params))
+    mock_response("translation" => self.find!(:help_center_translations, mock_translation_key).merge!(translation_params))
   end
 end
